requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
autobahn==19.11.2
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
Django==3.2.4
enum34==1.1.10
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kiteconnect==3.9.2
kombu==5.1.0
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
Pyrebase4==4.5.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
